Fix countdown bugs
When always harvesting is enabled, it's possible that the battery will
remain at its maximum energy capacity. So the countdown must be reset in
restore, not in power_on (which can be called repeatedly by is_active).

Also, if there is not enough energy to backup, we must continue to count
down. However, we should also check to see if the counter reaches 0. If
this is the case and we still haven't had enough energy to perform the
backup, we should power off.